Finished replacing player variables
